requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.9
django-colorfield==0.11.0
django-debug-toolbar==4.2.0
mutagen==1.47.0
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
